rename : entity folder to model folder
