+++ src/server/chat-service/src/main/java/com/example/chatservice/controller/RoomController.java
@@ -2,7 +2,7 @@ package com.example.chatservice.controller;
 
 import com.example.chatservice.dto.ChatDto;
 import com.example.chatservice.dto.RoomDto;
-import com.example.chatservice.entity.room.Room;
+import com.example.chatservice.model.room.Room;
 import com.example.chatservice.redis.RedisRoomRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
