+++ src/server/chat-service/src/main/java/com/example/chatservice/redis/RedisPublisher.java
@@ -1,6 +1,6 @@
 package com.example.chatservice.redis;
 
-import com.example.chatservice.entity.chat.Chat;
+import com.example.chatservice.model.chat.Chat;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
