+++ src/server/chat-service/src/main/java/com/example/chatservice/redis/RedisRoomRepository.java
@@ -6,7 +6,7 @@ import java.util.Map;
 
 import javax.annotation.PostConstruct;
 
-import com.example.chatservice.entity.room.Room;
+import com.example.chatservice.model.room.Room;
 
 import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.RedisTemplate;
